Changes 2-22
Changes after xh175 master by cmb69 commit c36c7f1d47f870c6a2c190af02c370f642437db6

+++ cmsimple/classes/Controller.php
@@ -545,7 +545,7 @@ class Controller
     {
         global $tx, $o;
 
-        $interval = 1000 * (ini_get('session.gc_maxlifetime') - 1);
+        $interval = 1000 * ((int)ini_get('session.gc_maxlifetime') - 1);
         $o .= <<<EOT
 <script type="text/javascript">
 if (document.cookie.indexOf('status=adm') == -1) {
